Amend rate-limit code and add more exception checking
The commit explicitly sets ratelimit_method to "wait" when initiating Mastodon (though this is the default anyway).
This setting allows mastodon.py to handle rate limits for us instead of having to put logic in ephemetoot to deal with it.

Also added some exception/error checking within the toot checking loop, allowing a second attempt for Mastodon API errors, adding a break after a KeyboardInterupt (Ctrl + C), and printing the errors in some cases to help with debugging.

Finally, there is now a 2 second delay between delete calls, just to be a bit nicer to the endpoint server.